+++ src/r_utils.h
@@ -36,7 +36,9 @@ static char guess_delim(
     end = source.size();
   }
 
-  auto nl = find_next_newline(
+  size_t nl;
+  newline_type nlt;
+  std::tie(nl, nlt) = find_next_newline(
       source,
       start,
       /* comment */ "",
@@ -46,8 +48,8 @@ static char guess_delim(
   while (nl > start && nl <= end && guess_max > 0) {
     auto str = std::string(source.data() + start, nl - start);
     lines.push_back(str);
-    start = nl + 1;
-    nl = find_next_newline(
+    start = nl + nlt == CRLF ? 2 : 1;
+    std::tie(nl, nlt) = find_next_newline(
         source,
         start,
         /* comment */ "",