+++ tests/testthat/test-vroom.R
@@ -814,3 +814,9 @@ test_that("vroom works if given col_names and col_types less than the number of 
   expect_equal(x[["x"]], "a")
   expect_equal(x[["X2"]], "b")
 })
+
+test_that("vroom works with CR line endings only", {
+  test_vroom(I("a,b\r1,2\r3,4\r"), delim = ",",
+    equals = tibble::tibble(a = c(1, 3), b = c(2, 4))
+  )
+})